+++ build_runner/lib/src/generate/phase.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:build/build.dart';
+import 'package:build_config/build_config.dart';
 import 'package:collection/collection.dart';
 
 import 'input_matcher.dart';
@@ -38,10 +39,9 @@ class BuildAction implements InputMatcher {
 
   /// Creates an [BuildAction] for a normal [Builder].
   ///
-  /// The build target is defined by [package] as well as [include] and
-  /// [exclude]. By default all sources in the target are used as primary inputs
-  /// to the builder, but it can be further filtered with [generateFor]. Glob
-  /// syntax is supported for [include], [exclude], and [generateFor].
+  /// The build target is defined by [package] as well as [targetSources]. By
+  /// default all sources in the target are used as primary inputs to the
+  /// builder, but it can be further filtered with [generateFor].
   ///
   /// [isOptional] specifies that a Builder may not be run unless some other
   /// Builder in a later phase attempts to read one of the potential outputs.
@@ -51,17 +51,15 @@ class BuildAction implements InputMatcher {
   factory BuildAction(
     Builder builder,
     String package, {
-    Iterable<String> include,
-    Iterable<String> exclude,
-    Iterable<String> generateFor,
+    InputSet targetSources,
+    InputSet generateFor,
     BuilderOptions builderOptions,
     bool isOptional,
     bool hideOutput,
   }) {
-    var inputs = new InputMatcher(include: include, exclude: exclude);
-    if (generateFor != null && generateFor.isNotEmpty) {
-      inputs = new InputMatcher.allOf(
-          [inputs, new InputMatcher(include: generateFor)]);
+    var inputs = new InputMatcher(targetSources ?? const InputSet());
+    if (generateFor != null) {
+      inputs = new InputMatcher.allOf([inputs, new InputMatcher(generateFor)]);
     }
     builderOptions ??= const BuilderOptions(const {});
     return new BuildAction._(package, builder, inputs, builderOptions,